+++ 1.21.4/Fabric/src/main/java/fuzs/puzzleslib/fabric/mixin/MobFabricMixin.java
@@ -9,9 +9,11 @@ import net.minecraft.nbt.CompoundTag;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.DifficultyInstance;
 import net.minecraft.world.entity.*;
+import net.minecraft.world.entity.ai.goal.GoalSelector;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.ServerLevelAccessor;
 import org.jetbrains.annotations.Nullable;
+import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
@@ -23,6 +25,12 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 @Mixin(Mob.class)
 abstract class MobFabricMixin extends LivingEntity implements SpawnTypeMob {
     @Shadow
+    @Final
+    public GoalSelector goalSelector;
+    @Shadow
+    @Final
+    public GoalSelector targetSelector;
+    @Shadow
     @Nullable
     private LivingEntity target;
     @Unique
@@ -31,6 +39,14 @@ abstract class MobFabricMixin extends LivingEntity implements SpawnTypeMob {
 
     protected MobFabricMixin(EntityType<? extends LivingEntity> entityType, Level level) {
         super(entityType, level);
+    }
+
+    @Inject(method = "<init>", at = @At("TAIL"))
+    protected void init(EntityType<? extends Mob> entityType, Level level, CallbackInfo callback) {
+        if (level instanceof ServerLevel) {
+            FabricLivingEvents.SETUP_MOB_GOALS.invoker()
+                    .onSetupMobGoals(Mob.class.cast(this), this.goalSelector, this.targetSelector);
+        }
     }
 
     @Inject(method = "finalizeSpawn", at = @At("TAIL"))
@@ -62,17 +78,14 @@ abstract class MobFabricMixin extends LivingEntity implements SpawnTypeMob {
     }
 
     @Inject(
-            method = "checkDespawn",
-            at = @At(
-                    value = "INVOKE",
-                    target = "Lnet/minecraft/world/level/Level;getNearestPlayer(Lnet/minecraft/world/entity/Entity;D)Lnet/minecraft/world/entity/player/Player;"
-            ),
-            cancellable = true
+            method = "checkDespawn", at = @At(
+            value = "INVOKE",
+            target = "Lnet/minecraft/world/level/Level;getNearestPlayer(Lnet/minecraft/world/entity/Entity;D)Lnet/minecraft/world/entity/player/Player;"
+    ), cancellable = true
     )
     public void checkDespawn(CallbackInfo callback) {
-        EventResult result = FabricLivingEvents.CHECK_MOB_DESPAWN.invoker().onCheckMobDespawn(Mob.class.cast(this),
-                (ServerLevel) this.level()
-        );
+        EventResult result = FabricLivingEvents.CHECK_MOB_DESPAWN.invoker()
+                .onCheckMobDespawn(Mob.class.cast(this), (ServerLevel) this.level());
         if (result.isInterrupt()) {
             if (result.getAsBoolean()) {
                 this.discard();