+++ 1.21.4/NeoForge/src/main/java/fuzs/puzzleslib/neoforge/impl/core/NeoForgeAbstractions.java
@@ -46,8 +46,7 @@ public final class NeoForgeAbstractions implements CommonAbstractions {
 
     @Override
     public boolean hasChannel(ServerPlayer serverPlayer, CustomPacketPayload.Type<?> type) {
-        // fake players will have no connection instance attached
-        return serverPlayer.connection != null && serverPlayer.connection.hasChannel(type);
+        return !serverPlayer.isFakePlayer() && serverPlayer.connection.hasChannel(type);
     }
 
     @Override